+++ cmd/kubeadm/app/cmd/upgrade/plan_test.go
@@ -22,7 +22,7 @@ import (
 	"testing"
 
 	kubeadmapi "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm"
-	"k8s.io/kubernetes/cmd/kubeadm/app/phases/upgrade"
+	"k8s.io/kubernetes/cmd/kubeadm/app/util/upgrade"
 )
 
 func TestSortedSliceFromStringIntMap(t *testing.T) {
